requirements.txt
Flask==2.3.2
torch==2.0.1
torchvision==0.15.2
numpy==1.25.2
Pillow==10.0.0
opencv-python==4.7.0.72
gunicorn
